move store managers into storeManagers folder
