+++ src/backend/storeManagers/gog/setup.ts
@@ -9,15 +9,15 @@ import {
 import { copySync } from 'fs-extra'
 import path from 'node:path'
 import { GameInfo, InstalledInfo } from 'common/types'
-import { checkWineBeforeLaunch, getShellPath, spawnAsync } from '../utils'
-import { GameConfig } from '../game_config'
-import { logError, logInfo, LogPrefix, logWarning } from '../logger/logger'
-import { isWindows } from '../constants'
+import { checkWineBeforeLaunch, getShellPath, spawnAsync } from '../../utils'
+import { GameConfig } from '../../game_config'
+import { logError, logInfo, LogPrefix, logWarning } from '../../logger/logger'
+import { isWindows } from '../../constants'
 import ini from 'ini'
-import { isOnline } from '../online_monitor'
-import { getWinePath, runWineCommand, verifyWinePrefix } from '../launcher'
-import { logFileLocation } from 'backend/gameManagerCommon/games'
-import { getGameInfo as getGogLibraryGameInfo } from 'backend/gog/library'
+import { isOnline } from '../../online_monitor'
+import { getWinePath, runWineCommand, verifyWinePrefix } from '../../launcher'
+import { logFileLocation } from 'backend/storeManagers/storeManagerCommon/games'
+import { getGameInfo as getGogLibraryGameInfo } from 'backend/storeManagers/gog/library'
 const nonNativePathSeparator = path.sep === '/' ? '\\' : '/'
 
 /**
