+++ src/backend/storeManagers/gog/user.ts
@@ -1,9 +1,9 @@
 import axios from 'axios'
-import { logError, logInfo, LogPrefix, logWarning } from '../logger/logger'
+import { logError, logInfo, LogPrefix, logWarning } from '../../logger/logger'
 import { GOGLoginData } from 'common/types'
-import { configStore, libraryStore } from '../gog/electronStores'
-import { errorHandler } from '../utils'
-import { isOnline } from '../online_monitor'
+import { configStore, libraryStore } from './electronStores'
+import { errorHandler } from '../../utils'
+import { isOnline } from '../../online_monitor'
 import { UserData } from 'common/types/gog'
 
 const gogAuthenticateUrl =
