+++ src/backend/storeManagers/hyperplay/utils.ts
