+++ src/backend/storeManagers/legendary/electronStores.ts
@@ -1,4 +1,4 @@
-import { TypeCheckedStoreBackend } from '../electron_store'
+import { TypeCheckedStoreBackend } from '../../electron_store'
 
 const installStore = new TypeCheckedStoreBackend('legendaryInstallInfo', {
   cwd: 'lib-cache',
