+++ src/backend/storeManagers/legendary/eos_overlay/ipc_handler.ts
