+++ src/backend/storeManagers/sideload/electronStores.ts
@@ -1,4 +1,4 @@
-import { TypeCheckedStoreBackend } from '../electron_store'
+import { TypeCheckedStoreBackend } from '../../electron_store'
 
 export const libraryStore = new TypeCheckedStoreBackend('sideloadedStore', {
   cwd: 'sideload_apps',
