+++ src/backend/storeManagers/sideload/library.ts
